+++ examples/charting/src/main.fixture.ts
