+++ examples/charting/src/scene.test.ts
@@ -15,7 +15,7 @@ import { describe, test } from 'vitest'
 import { RadioGroup } from '@foldkit/ui'
 
 import { SyncChart } from './command'
-import { loadingModel, readyModel, sampleTelemetry } from './main.fixtures'
+import { loadingModel, readyModel, sampleTelemetry } from './main.fixture'
 import { Message } from './message'
 import { TelemetryAsyncData } from './model'
 import { update } from './update'
