+++ examples/map/src/main.fixture.ts
