+++ examples/weather/src/main.fixture.ts
